// src/app/model/report-approval.model.ts
export class ReportApproval {
    id: number = 0;
    userId: number = 0;
    userName: string = "";
    reportMonth: number = 0;
    reportYear: number = 0;
    isApproved: boolean = false;
    notes: string = "";
    createdBy: number = 0;
    createdData: any;
    updatedBy: number = 0;
    updatedDate: any;
    isDeleted: boolean = false;
}